+++ src/cli/commands/init/init.ts
@@ -15,7 +15,7 @@ import { collectTelemetryEvent } from "../../../core/telemetry/utils";
 import { TELEMETRY_EVENTS } from "../../../core/constants";
 
 export async function init(options: SWACLIConfig, showHints: boolean = true) {
-  const start = new Date().getTime();
+  const cmdStartTime = new Date().getTime();
   const configFilePath = options.config!;
   const disablePrompts = options.yes ?? false;
   const outputFolder = process.cwd();
@@ -133,18 +133,11 @@ export async function init(options: SWACLIConfig, showHints: boolean = true) {
 
     logger.log(`- Use ${chalk.cyan("swa deploy")} to deploy your app to Azure.\n`);
   }
-  const end = new Date().getTime();
-  const appFramework = projectConfig?.name?.split(", with")[0]!;
-  if (appFramework) {
-    await collectTelemetryEvent(TELEMETRY_EVENTS.Init, {
-      appFramework: appFramework,
-      duration: (end - start).toString(),
-    });
-  } else {
-    await collectTelemetryEvent(TELEMETRY_EVENTS.Init, {
-      duration: (end - start).toString(),
-    });
-  }
+  const cmdEndTime = new Date().getTime();
+  await collectTelemetryEvent(TELEMETRY_EVENTS.Init, {
+    appFramework: projectConfig?.name?.split(", with")[0]!,
+    duration: (cmdEndTime - cmdStartTime).toString(),
+  });
 }
 
 function convertToCliConfig(config: FrameworkConfig): SWACLIConfig {
